refactor: move and rename NextLessonCard
